Changed connection queries to return 0 instead of null when no connections are available Changed SQL connection query to return all DBs and connection #'s even if no connections are made to a DB Added int to the list of numeric types that can be converted to/from a Count/Value metric

+++ src/NewRelic.Microsoft.SqlServer.Plugin/MetricMapper.cs
@@ -9,7 +9,7 @@ namespace NewRelic.Microsoft.SqlServer.Plugin
 {
 	public class MetricMapper
 	{
-		private static readonly Type[] _NumericMappingTypes = new[] {typeof (long), typeof (byte), typeof (short), typeof (decimal), typeof(int)};
+		private static readonly Type[] _NumericMappingTypes = new[] {typeof (long), typeof(int), typeof (byte), typeof (short), typeof (decimal),};
 		private static readonly Type[] _NumericMetricTypes = new[] {typeof (decimal), typeof (int)};
 		private readonly MapAction _metricSetter;
 		private readonly PropertyInfo _propertyInfo;
